requirements.txt
torch>=2.1.0
torchvision>=0.16.0
torch-geometric>=2.4.0
transformers>=4.35.0
diffusers>=0.23.0
openai>=1.3.0
numpy>=1.24.0
scipy>=1.11.0
scikit-learn>=1.3.0
psutil>=5.9.0
pyyaml>=6.0
gradio>=4.0.0  # For dashboard
pytest>=7.4.0